split decorator.py from udfgenerator.py
